fix: :lipstick: footer-update
Still not 100% like I want it to function, but way better than before. Currently the footer is sticking to the bottom of the users viewport, instead of it floating in the middle of bum-f*ck nowhere on the middle of the page. Background is added when screen is around phone-size, although tweaks will need to be made, for it to look even better.

Idk if anything is broken or not. We need someone to implement routing for different pages other than App.tsx + I need all the different JSON-structures that the API returns, to create mock-data for testing purposes.