fix: :bug: Crafting-fix, leaderboard-glitch fix, styling changes, text-changes, clarification of function (Tooltips or rewritten text), change version-number behavior
Crafting recipes now show no matter what order item1 and item2 are. - Changed styling after feedback from Pixelguy, to clearify if a button is enabled/disabled (Reverted colors on non-hover and hover on settings-buttons. Reverted colors on "Buy" in shop and added an underline) - Added yellow text-color to users QP amount in the shop to highlight their current balance more. - Changed behavior of the version-number in the footer, it's no longer clickable. - Clearified what "Autocomplete" means. - Clearified, with a tooltip, what the "Reset"-button on the Border-selection does. - Changed "Go to your profile and click on 'Complete Challenge'" to "Go to the 'Challenges'-page and click on 'Complete Challenge'" instead.

Inventory & Crafting might break, I have not fully tested it since I lack the items to do so.