feat: :sparkles: Type-corrections + addition. Updated styling
Changed "Banner"-type from string to boolean. Added "admin: boolean" and "banned: boolean" to the UserInfo-interface. Updated styling. Footer *should* now correctly place itself at the bottom of the page

Potentially breaking user logins, given that the advanced endpoint is lacking "admin" and "banned" data. I've yet to get an error on logging in, but @ me on Discord if this happens.